back/requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.2
python-decouple==3.8
Pillow==10.1.0
openpyxl==3.11.0
pandas==2.1.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
